Update backend configuration: add postinstall script for Prisma generation, set Node engine requirement, and refactor index.ts to streamline server startup process by importing app.js and removing unnecessary middleware.
