+++ backend/src/index.ts
@@ -1,78 +1,8 @@
-import express from 'express';
-import cors from 'cors';
-import helmet from 'helmet';
-import morgan from 'morgan';
-import rateLimit from 'express-rate-limit';
-import dotenv from 'dotenv';
+import app from './app.js';
 
-import authRoutes from './routes/auth.js';
-import dashboardRoutes from './routes/dashboard.js';
-import inventoryRoutes from './routes/inventory.js';
-import rankingsRoutes from './routes/rankings.js';
-import { errorHandler } from './middleware/errorHandler.js';
-import { notFound } from './middleware/notFound.js';
-import { ensureAdminExists } from './lib/bootstrap.js';
-
-// Load environment variables
-dotenv.config();
-
-// Ensure a fixed admin exists on startup (before serving requests)
-await ensureAdminExists().catch((err) => {
-  console.error('Admin bootstrap failed:', err);
-});
-
-const app = express();
 const PORT = process.env.PORT || 4000;
 
-// Security middleware
-app.use(helmet());
-
-// CORS configuration
-app.use(cors({
-  origin: process.env.FRONTEND_URL || 'http://localhost:5173',
-  credentials: true,
-  methods: ['GET', 'POST', 'PUT', 'DELETE', 'PATCH', 'OPTIONS'],
-  allowedHeaders: ['Content-Type', 'Authorization']
-}));
-
-// Rate limiting
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // 15 minutes
-  max: 100, // limit each IP to 100 requests per windowMs
-  message: 'Too many requests from this IP, please try again later.',
-  standardHeaders: true,
-  legacyHeaders: false,
-});
-app.use(limiter);
-
-// Logging
-app.use(morgan('combined'));
-
-// Body parsing middleware
-app.use(express.json({ limit: '10mb' }));
-app.use(express.urlencoded({ extended: true, limit: '10mb' }));
-
-// Health check endpoint
-app.get('/health', (req, res) => {
-  res.status(200).json({
-    status: 'OK',
-    timestamp: new Date().toISOString(),
-    uptime: process.uptime(),
-    environment: process.env.NODE_ENV || 'development'
-  });
-});
-
-// API routes
-app.use('/api/auth', authRoutes);
-app.use('/api/dashboard', dashboardRoutes);
-app.use('/api/inventory', inventoryRoutes);
-app.use('/api/rankings', rankingsRoutes);
-
-// Error handling middleware
-app.use(notFound);
-app.use(errorHandler);
-
-// Start server
+// Start server (for local / traditional hosting). On Vercel, this file isn't used.
 app.listen(PORT, () => {
   console.log(`🚀 Server running on port ${PORT}`);
   console.log(`📊 Environment: ${process.env.NODE_ENV || 'development'}`);
